gcp-function-source/requirements.txt
# Function dependencies, for example:
# package>=version
requests
scikit-learn
sklearn
google-cloud-storage
numpy
pandas